Wire up Add Object and Mark Empty buttons
