+++ src/features/loupe/ImageReviewToolbar.jsx
@@ -1,42 +1,37 @@
 import { styled } from '../../theme/stitches.config.js';
-import { ToolbarRoot, ToolbarToggleItem, ToolbarSeparator, ToolbarLink, ToolbarButton, ToolbarIconButton } from '../../components/Toolbar.jsx';
-import { ToggleGroup } from '@radix-ui/react-toolbar';
+import { ToolbarRoot, ToolbarSeparator, ToolbarButton, ToolbarIconButton } from '../../components/Toolbar.jsx';
 import {
   PlusIcon,
   ValueNoneIcon,
   CheckIcon,
   Cross2Icon,
-  StrikethroughIcon,
-  TextAlignLeftIcon,
-  TextAlignCenterIcon,
-  TextAlignRightIcon,
-  FontBoldIcon,
-  FontItalicIcon,
 } from '@radix-ui/react-icons';
 
 const StyledToolbarRoot = styled(ToolbarRoot, {
   marginTop: '$2'
 })
 
-const ImageReviewToolbar = () => (
-  <StyledToolbarRoot aria-label="Formatting options">
-    <ToolbarIconButton>
-      <CheckIcon /> Validate
-    </ToolbarIconButton>
-    <ToolbarSeparator />
-    <ToolbarIconButton>
-      <Cross2Icon /> Invalidate
-    </ToolbarIconButton>
-    <ToolbarSeparator />
-    <ToolbarIconButton>
-      <ValueNoneIcon /> Mark empty
-    </ToolbarIconButton>
-    <ToolbarSeparator />
-    <ToolbarIconButton>
-      <PlusIcon /> Add object
-    </ToolbarIconButton>
-    {/* <ToolbarButton css={{ marginLeft: 'auto' }}>Share</ToolbarButton> */}
-  </StyledToolbarRoot>
-);
+const ImageReviewToolbar = ({ handleMarkEmptyButtonClick, handleAddObjectButtonClick }) => {
+  return (
+    <StyledToolbarRoot aria-label="Formatting options">
+      <ToolbarIconButton>
+        <CheckIcon /> Validate
+      </ToolbarIconButton>
+      <ToolbarSeparator />
+      <ToolbarIconButton>
+        <Cross2Icon /> Invalidate
+      </ToolbarIconButton>
+      <ToolbarSeparator />
+      <ToolbarIconButton onClick={handleMarkEmptyButtonClick}>
+        <ValueNoneIcon /> Mark empty
+      </ToolbarIconButton>
+      <ToolbarSeparator />
+      <ToolbarIconButton onClick={handleAddObjectButtonClick}>
+        <PlusIcon /> Add object
+      </ToolbarIconButton>
+      {/* <ToolbarButton css={{ marginLeft: 'auto' }}>Share</ToolbarButton> */}
+    </StyledToolbarRoot>
+  );
+};
 
 export default ImageReviewToolbar;
